Fix environment formatting, default no NFS environments

+++ new_jupyter_sciencedata/before-notebook.d/envdir.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
-[[ -d /mnt/sciencedata/jupyter/.conda ]] || mkdir -p /mnt/sciencedata/jupyter/.conda
-rm -rf /home/sciencedata/.conda
-ln -s /mnt/sciencedata/jupyter/.conda /home/sciencedata/.conda
+if [[ $NFS_ENVIRONMENTS ]]; then
+	[[ -d /mnt/sciencedata/jupyter/.conda ]] || mkdir -p /mnt/sciencedata/jupyter/.conda
+	rm -rf /home/sciencedata/.conda
+	ln -s /mnt/sciencedata/jupyter/.conda /home/sciencedata/.conda
+fi
